refactor(inferno-component): Adjust imports and types
