+++ packages/inferno-component/inferno-component.js
@@ -1,2 +1,2 @@
-module.exports = require('./../inferno/dist/inferno-component.node');
+module.exports = require('./dist/inferno-component.node');
 module.exports.default = module.exports;
